refactor: styles and format
