+++ docs/.vitepress/config.ts
@@ -4,6 +4,7 @@ export default defineConfig({
   lang: 'en-US',
   title: 'VitePress',
   description: 'Vite & Vue powered static site generator.',
+
   lastUpdated: true,
 
   themeConfig: {
