--- src/client/theme-default/components/VPLastUpdated.vue
@@ -1,26 +1,0 @@
-<script setup lang="ts">
-import { ref, onMounted, watchEffect } from 'vue'
-import { useData } from 'vitepress'
-
-const { theme, page } = useData()
-const date = new Date(page.value.lastUpdated!)
-const isoDatetime = date.toISOString()
-const datetime = ref('')
-
-onMounted(() => {
-  watchEffect(() => {
-    // locale string might be different based on end user
-    // and will lead to potential hydration mismatch if calculated at build time
-    datetime.value = date.toLocaleString(window.navigator.language)
-  })
-})
-</script>
-
-<template>
-  <div>
-    {{ theme.lastUpdated ?? 'Last Updated' }}:
-    <time :datatime="isoDatetime">
-      {{ datetime }}
-    </time>
-  </div>
-</template>
